24/06/19 불필요한 코드 삭제 및 정리, Api.js 분리
